refactor(gatsby-source-contentful): move deprecated gatsby-image code to separate file
