+++ packages/gatsby-source-contentful/src/__tests__/image-helpers.js
@@ -1,0 +1,58 @@
+const { createUrl, getBase64Image } = require(`../image-helpers`)
+
+describe(`contentful extend node type`, () => {
+  describe(`createUrl`, () => {
+    it(`allows you to create URls`, () => {
+      expect(
+        createUrl(`//images.contentful.com/dsf/bl.jpg`, { width: 100 })
+      ).toMatchSnapshot()
+    })
+    it(`ignores options it doesn't understand`, () => {
+      expect(
+        createUrl(`//images.contentful.com/dsf/bl.jpg`, { happiness: 100 })
+      ).toMatchSnapshot()
+    })
+  })
+
+  describe(`getBase64Image`, () => {
+    const imageProps = {
+      aspectRatio: 4.8698224852071,
+      baseUrl: `//images.ctfassets.net/k8iqpp6u0ior/3ljGfnpegOnBTFGhV07iC1/94257340bda15ad4ca8462da3a8afa07/347966-contentful-logo-wordmark-dark__1_-4cd185-original-1582664935__1_.png`,
+      width: 200,
+      height: 41,
+      image: {
+        contentful_id: `3ljGfnpegOnBTFGhV07iC1`,
+        spaceId: `k8iqpp6u0ior`,
+        createdAt: `2021-03-22T10:10:34.647Z`,
+        updatedAt: `2021-03-22T10:10:34.647Z`,
+        file: { contentType: `image/png` },
+        title: `Contentful Logo PNG`,
+        description: ``,
+        node_locale: `en-US`,
+      },
+      options: {
+        width: 200,
+        quality: 50,
+        toFormat: ``,
+        cropFocus: null,
+        cornerRadius: 0,
+        background: null,
+      },
+    }
+    test(`keeps image format`, async () => {
+      const result = await getBase64Image(imageProps)
+      expect(result).toMatch(
+        `data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAECAMAAAC5ge+kAAAAllBMVEUAAABHl745rOE7tOc7tOcqMDkqMDkqMDkqMDnfzG9Pm7o7tOc7tOcqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDn4wF/eXWDtXGjtXGgqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDkqMDnbVmDpWGbtXGjtXGh1tTylAAAAMnRSTlMATd3gVSUjTCDgHRIscF+MeqB8qpqbk4ienYAxr+AeEipyZI9/aW+No4WJeWuuTdzgVnu3oiUAAAAJcEhZcwAACxIAAAsSAdLdfvwAAAAHdElNRQflCAUMNjFcK/NJAAAAMklEQVQI12NkBII/DCDA+htIsDEy/mBj4WDEBCwiyLwnIpyMjL/ZWASB7PMMMPAZTAIALlUHKTqI1/MAAAAASUVORK5CYII=`
+      )
+    })
+    test(`uses given image format`, async () => {
+      const result = await getBase64Image({
+        ...imageProps,
+        options: { ...imageProps.options, toFormat: `jpg` },
+      })
+      expect(result).toMatch(
+        `data:image/jpg;base64,/9j/4AAQSkZJRgABAQIAHAAcAAD/2wBDABALDA4MChAODQ4SERATGCgaGBYWGDEjJR0oOjM9PDkzODdASFxOQERXRTc4UG1RV19iZ2hnPk1xeXBkeFxlZ2P/2wBDARESEhgVGC8aGi9jQjhCY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2P/wAARCAAEABQDASIAAhEBAxEB/8QAFwABAAMAAAAAAAAAAAAAAAAAAAIDBv/EACQQAAIBAgQHAQAAAAAAAAAAAAECAAMRBBITJAUUFSFBUWHB/8QAFQEBAQAAAAAAAAAAAAAAAAAAAgH/xAAXEQEBAQEAAAAAAAAAAAAAAAABAAIx/9oADAMBAAIRAxEAPwDV4NObWqM70dOoVvROUt9Psy7pYud5jO/jWiJM8lsDSFB+Do+Xe4xQosAtW35ERFC//9k=`
+      )
+    })
+  })
+})
